Move wolfiRebuildAllBaseImages
